Fix build errors/warnings across all projects under -Wall -Werror -Wextra
- ft_memchr: fix void* pointer arithmetic/indexing (GNU extension,
  rejected by strict gcc) in printf, pipex, push_swap, minishell, cube3D
- printunsigneddecimal: remove dead `if (c < 0)` check on unsigned int
  (always false, -Werror=type-limits) in fdf, pipex, push_swap
- pipex/push_swap Makefiles: fix -lft linked before objects (undefined
  reference under GNU ld), add missing libft build step and .PHONY
- push_swap bonus checker: fix uninitialized t_data.a/b read before
  stack_init (segfault), fix get_next_line output never trimmed of
  trailing newline before comparing against move names (always "Error !")
- philo: fix 32-bit int overflow in get_time()/sleep_time() causing
  garbage timestamps and busy-spin (sleep never actually delayed);
  death message now prints relative time instead of absolute epoch;
  added fflush(stdout) so piped output isn't fully buffered
- minishell Makefile: remove hardcoded teammate-specific Homebrew path,
  auto-detect readline via `brew --prefix` on macOS / system -lreadline
  on Linux
- CPP09/ex02 PmergeMe: fix std::find() called on queue::front()/back()
  values instead of iterators (compile error); note FordJohnson sort
  itself is still a stub, not implemented
- untrack stale build binaries removed by .gitignore

Known issue (not fixed, deep algorithmic bug): push_swap sort produces
incorrect KO results for n>5; execute_cheapest() in algorythm.c mixes
inconsistent cost formulas (|a|+|b| vs |a+b|).

+++ philo/header/structures.h
@@ -18,7 +18,7 @@
 typedef struct s_philo
 {
 	int				id;
-	int				last_meal;
+	long			last_meal;
 	pthread_t		thread;
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	*right_fork;
@@ -33,7 +33,7 @@ typedef struct s_data
 	int				t_eat;
 	int				t_sleep;
 	int				nb_eat;
-	int				start;
+	long			start;
 	int				run;
 	pthread_mutex_t	print;
 	pthread_mutex_t	run_lock;
